+++ pyscriptjs/src/pyexec.ts
@@ -12,7 +12,7 @@ export async function pyExec(
     outElem: HTMLElement,
 ): Promise<{ result: any }> {
     ensureUniqueId(outElem);
-    if (await interpreter._remote.pyscript_py.uses_top_level_await(pysrc)) {
+    if (await interpreter._remote.pyscript_internal.uses_top_level_await(pysrc)) {
         const err = new UserError(
             ErrorCode.TOP_LEVEL_AWAIT,
             'The use of top-level "await", "async for", and ' +
