+++ pyscriptjs/tests/integration/test_02_display.py
@@ -330,8 +330,8 @@ class TestOutput(PyScriptTest):
             """
             <py-script>
                 print('print from python')
-                console.log('print from js')
-                console.error('error from js');
+                js.console.log('print from js')
+                js.console.error('error from js');
             </py-script>
         """
         )
@@ -348,8 +348,8 @@ class TestOutput(PyScriptTest):
             <py-script>
                 display('this goes to the DOM')
                 print('print from python')
-                console.log('print from js')
-                console.error('error from js');
+                js.console.log('print from js')
+                js.console.error('error from js');
             </py-script>
         """
         )
