+++ pyscriptjs/tests/py-unit/_pyscript_js.py
@@ -1,0 +1,4 @@
+from unittest.mock import Mock
+
+showWarning = Mock()
+define_custom_element = Mock()
