+++ pyscriptjs/tests/py-unit/js.py
@@ -4,3 +4,4 @@ from unittest.mock import Mock
 document = Mock()
 console = Mock()
 setTimeout = Mock()
+Object = Mock()
